fix: move defaultDirFor to shared paths.go for Windows compilation
